refactor: project items list
